Give the tensor type domain a file per concept
A mod.rs is a door, not a definition site, and tensor/mod.rs had grown to
1,100 lines of definitions. Move the type side out of it: dtype.rs for the
element type, tensor_type.rs for a shape of per-axis sizes and the type that
pairs one with a dtype, and error.rs for the domain's error set.

error.rs rather than folding TensorError into the tensor value: of its four
variants the value side constructs only DTypeMismatch, and the other three
belong to broadcasting and to the type-level rules, so no single concept owns
the enum. fmt_shape becomes pub(super) for it, and three doc links become
qualified now that Tensor and rules are no longer in the same scope.

Pure movement: the tests move with the items they cover and the test count is
unchanged.
